requirements.txt
google-genai>=0.6.0
chromadb>=0.5.0
streamlit>=1.35.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
